refactor dashboard with components and composables
